feature(Monday): Refactoring Code
